ipbilling/opbilling

+++ Libraries/HIMS.Data/DTO/IPPatient/IPAddchargesListDto.cs
@@ -53,8 +53,8 @@ namespace HIMS.Data.DTO.IPPatient
         public string? ServiceCode { get; set; }
         public bool? IsInclusionExclusion { get; set; }
         public long UnitId { get; set; }
-        public long? WardId { get; set; }
-        public long? BedId { get; set; }
+        public int? WardId { get; set; }
+        public int? BedId { get; set; }
     }
 
 }
